requirements.txt
nextcord==2.5.0
python-dotenv==0.21.0
# aiohttp==3.8.4